(MODULES-10950) Infer application name from run mode
Fix a regression introduced by MODULES-7397[1] which incorrectly assumed
that the application name can be queried through
`Puppet::Application.name`, causing lenses that are pluginsynced not to
be loaded.

Since we only need to find out whether or not we're running as part of
`puppet agent`, it should be enough to use `Puppet.run_mode.name`, which
returns `:agent` in this case, and `:user` otherwise.

[1] https://github.com/puppetlabs/puppetlabs-augeas_core/pull/27