(maint) Fix up pre-suite for beaker tests
Solaris packing was updated that requires a newer cert in the keychain of the
VMs that we use. However for the older version of Solaris the cert is not on the
keychain. This PR calls a beaker-puppet script that will add the required cert
to the keychain of the VM.
Also follow redirects when downloading agent.