Remove unnecessary release prep gems
Prior to this commit, we added the concurrent-ruby gem in c2e7fa2 to
compensate for issues relating to the Docker image used for release
preparation.

Since then, we have switched to piggy-backing off of the CAT team's
release prep GitHub Action (see 65a59a3). That action stopped using the
problematic Docker image in mid-2023
(puppetlabs/cat-github-actions@4a83f93), so we no longer need to manage
concurrent-ruby.

This commit also removes the github_changelog_generator gem, as it has been
superceded by the gh-changelog tool to generate changelogs for releases.